Finished FETCH and mostly finished DECODE.
FETCH is pretty simple, so I'm pretty sure I got it right.

DECODE, on the other hand, is a little more difficult. I've finished everything except setting the DepStallSignal. It's pretty complicated to do the way the framework requires and I have a few questions about it. I'm also not sure when Branches should be handled (it's probably deeper in the code).